% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{count_infected_neighbors}
\alias{count_infected_neighbors}
\title{Count the number cells that are infected near a specific cell}
\usage{
count_infected_neighbors(x, i, j)
}
\arguments{
\item{x}{Input matrix}

\item{i}{Row coordinate}

\item{j}{Column coordinate}
}
\value{
Count
}
\description{
Count the number cells that are infected near a specific cell
}
\examples{
x <- create_matrix(5,5)
count_infected_neighbors(x, 2, 2)
}
