% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{create_cntr_matrix}
\alias{create_cntr_matrix}
\title{Create a matrix of any size with only a single cell in the center starts as infected}
\usage{
create_cntr_matrix(row, col)
}
\arguments{
\item{row}{Number of rows for matrix}

\item{col}{Number of columns for matrix}
}
\value{
Infection Matrix
}
\description{
Create a matrix of any size with only a single cell in the center starts as infected
}
\examples{
create_cntr_matrix(10,10)
}
