% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{create_crnr_matrix}
\alias{create_crnr_matrix}
\title{Create corner starting matrix}
\usage{
create_crnr_matrix(row = 20, col = 30)
}
\arguments{
\item{row}{Number of rows in matrix}

\item{col}{Number of columns in matrix}
}
\value{
SIRmatrix
}
\description{
Create corner starting matrix
}
\examples{
create_crnr_matrix(20, 30)
}
