% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{dis_plot}
\alias{dis_plot}
\title{Plot the matrix}
\usage{
dis_plot(input)
}
\arguments{
\item{input}{Input matrix}
}
\value{
Matrix Image
}
\description{
Plot the matrix
}
\examples{
x <- create_matrix(5,5)

dis_plot(x)
}
