% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{infect_step}
\alias{infect_step}
\title{Single time step in simulation}
\usage{
infect_step(
  prob_infect = 0.25,
  input,
  auto_immunity = FALSE,
  imm_prob = 0.5,
  allow_death = FALSE,
  fat_prob = 0.15
)
}
\arguments{
\item{prob_infect}{Probability of infection spreading}

\item{input}{Input matrix}

\item{auto_immunity}{Enable auto immunity after infection}

\item{imm_prob}{Probability of getting immunity}

\item{allow_death}{Enable death}

\item{fat_prob}{Probability of fatality}
}
\value{
new_inf_matrix
}
\description{
Single time step in simulation
}
\examples{
x <- create_matrix(5,5)

infect_step(0.25,x)
infect_step(0.25,x,FALSE,0.25)
infect_step(0.25,x,FALSE,0.3,FALSE)
infect_step(0.25,x,FALSE,0.25,TRUE,0.15)
infect_step(0.25,x,TRUE,1.0,TRUE,0.09)
}
