% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{infection_step}
\alias{infection_step}
\title{Single step in simulation}
\usage{
infection_step(input = init_matrix(5, 5), inf_prob = 0.5)
}
\arguments{
\item{input}{Input SIR Matrix}

\item{inf_prob}{Probability of spreading infection}
}
\value{
Updated SIR Matrix after infection spread during single step
}
\description{
Single step in simulation
}
\examples{
infection_step(, 0.5)
}
