% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{init_matrix}
\alias{init_matrix}
\title{Create infection matrix}
\usage{
init_matrix(row, col, start_prop = 0.125)
}
\arguments{
\item{row}{Number of rows in matrix}

\item{col}{Number of columns in matrix}

\item{start_prop}{Probability of cells initially infected}
}
\value{
SIR Matrix
}
\description{
Create infection matrix
}
\examples{
init_matrix(15, 15, 0.125)
}
