% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{multiple_sims}
\alias{multiple_sims}
\title{Run multiple simulations}
\usage{
multiple_sims(prob, runs, row, col)
}
\arguments{
\item{prob}{Probability of spreading infection to other cells}

\item{runs}{Number of simulation runs}

\item{row}{Number of rows in matrix}

\item{col}{Number of columns in matrix}
}
\value{
List with each simulation statistics and final infection matrix
}
\description{
Run multiple simulations
}
\examples{
multiple_sims(0.5, 1000, 15, 15)
}
