% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{pad}
\alias{pad}
\title{Pad the input matrix so we don't have to worry about the edges}
\usage{
pad(input)
}
\arguments{
\item{input}{Input matrix}
}
\value{
input_pad
}
\description{
Pad the input matrix so we don't have to worry about the edges
}
\examples{
x <- create_matrix(5,5)

pad(x)
}
