% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{simulation}
\alias{simulation}
\title{Run simulation with different infection probabilities}
\usage{
simulation(prob, step = 0.1, row, col)
}
\arguments{
\item{prob}{Probability of spreading infection}

\item{step}{Value to increment probability range by}

\item{row}{Number of rows in matrix}

\item{col}{Number of columns in matrix}
}
\value{
Dataframe with statistics about all simulations run
}
\description{
Run simulation with different infection probabilities
}
\examples{
simulation(0.5, 1000, 20, 30)
}
