% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{create_crnr_matrix}
\alias{create_crnr_matrix}
\title{Create a matrix of any size with only the four corners starting as infected cells}
\usage{
create_crnr_matrix(row, col)
}
\arguments{
\item{row}{Number of rows for matrix}

\item{col}{Number of columns for matrix}
}
\value{
Infection Matrix
}
\description{
Create a matrix of any size with only the four corners starting as infected cells
}
\examples{
create_crnr_matrix(10,10)
}
