% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{create_matrix}
\alias{create_matrix}
\title{Create a matrix of any size with a random sample of infected starting cells}
\usage{
create_matrix(row, col, start_inf = 0.125)
}
\arguments{
\item{row}{Number of rows for matrix}

\item{col}{Number of columns for matrix}

\item{start_inf}{Probability that a cell will start as infected}
}
\value{
Infection Matrix
}
\description{
Create a matrix of any size with a random sample of infected starting cells
}
\examples{
create_matrix(5,5)
create_matrix(15,15,0.25)
}
