% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{create_random_matrix}
\alias{create_random_matrix}
\title{Create a matrix of any size with random infected starting cells}
\usage{
create_random_matrix(row, col, start_infected = 1, seed = NULL)
}
\arguments{
\item{row}{Number of rows for matrix}

\item{col}{Number of columns for matrix}

\item{start_infected}{Exact number of randomly selected cells that start infected}

\item{seed}{Optional random seed for reproducible sampling}
}
\value{
Infection Matrix
}
\description{
Create a matrix of any size with random infected starting cells
}
\examples{
create_random_matrix(5, 5)
create_random_matrix(15, 15, 10)
}
