% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{multiple_run_heatmap}
\alias{multiple_run_heatmap}
\title{Run a given simulation multiple times to summarize which cells were previously infected}
\usage{
multiple_run_heatmap(
  prob_infect = 0.25,
  input_matrix,
  runs = 10,
  plot = FALSE,
  seed = NULL
)
}
\arguments{
\item{prob_infect}{Probability of infection spreading}

\item{input_matrix}{Input matrix}

\item{runs}{Number of simulation runs}

\item{plot}{Plot the heatmap}

\item{seed}{Optional random seed for reproducible simulations}
}
\value{
Matrix of recovered counts by cell
}
\description{
Run a given simulation multiple times to summarize which cells were previously infected
}
\examples{
x <- create_matrix(5,5)

multiple_run_heatmap(0.25,x,10)
}
