% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_inf_seq}
\alias{simulate_inf_seq}
\title{Run a given simulation multiple times using a sequence of increasing infection probabilities.}
\usage{
simulate_inf_seq(
  input_matrix,
  step = 0.1,
  seed = NULL
)
}
\arguments{
\item{input_matrix}{Input matrix}

\item{step}{Value to step by in sequence of infection probabilities}

\item{seed}{Optional random seed for reproducible simulations}
}
\value{
Results Dataframe
}
\description{
Run a given simulation multiple times using a sequence of increasing infection probabilities.
}
\examples{
x <- create_matrix(5,5)

simulate_inf_seq(x,0.1)
}
