% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{simulate_inf_seq}
\alias{simulate_inf_seq}
\title{Run a given simulation multiple times using a sequence of increasing infection probabilities.}
\usage{
simulate_inf_seq(
  input_matrix,
  step = 0.1,
  auto_immunity = FALSE,
  imm_prob = 0.5,
  allow_death = FALSE,
  fat_prob = 0.15
)
}
\arguments{
\item{input_matrix}{Input matrix}

\item{step}{Value to step by in sequence of infection probabilities}

\item{auto_immunity}{Enable auto immunity after infection}

\item{imm_prob}{Probability of getting immunity}

\item{allow_death}{Enable death}

\item{fat_prob}{Probability of fatality}
}
\value{
Results Dataframe
}
\description{
Run a given simulation multiple times using a sequence of increasing infection probabilities.
}
\examples{
x <- create_matrix(5,5)

simulate_inf_seq(x,0.1,TRUE)
simulate_inf_seq(x,0.2,FALSE,0.35,FALSE)
simulate_inf_seq(x,0.1,FALSE,0.35,TRUE,0.12)
}
