% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_many_runs}
\alias{simulate_many_runs}
\title{Run a given simulation multiple times and get the average steps and average proportion of cells infected.}
\usage{
simulate_many_runs(
  prob_infect = 0.25,
  input_matrix,
  runs = 10,
  seed = NULL
)
}
\arguments{
\item{prob_infect}{Probability of infection spreading}

\item{input_matrix}{Input matrix}

\item{runs}{Number of simulation runs}

\item{seed}{Optional random seed for reproducible simulations}
}
\value{
Results Dataframe
}
\description{
Run a given simulation multiple times and get the average steps and average proportion of cells infected.
}
\examples{
x <- create_matrix(5,5)

simulate_many_runs(0.25,x,10)
}
