% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{simulate_many_runs}
\alias{simulate_many_runs}
\title{Run a given simulation multiple times and get the average steps and average proportion of cells infected.}
\usage{
simulate_many_runs(
  prob_infect = 0.25,
  input_matrix,
  runs = 10,
  auto_immunity = FALSE,
  imm_prob = 0.5,
  allow_death = FALSE,
  fat_prob = 0.15
)
}
\arguments{
\item{prob_infect}{Probability of infection spreading}

\item{input_matrix}{Input matrix}

\item{runs}{Number of simulation runs}

\item{auto_immunity}{Enable auto immunity after infection}

\item{imm_prob}{Probability of getting immunity}

\item{allow_death}{Enable death}

\item{fat_prob}{Probability of fatality}
}
\value{
Results Dataframe
}
\description{
Run a given simulation multiple times and get the average steps and average proportion of cells infected.
}
\examples{
x <- create_matrix(5,5)

simulate_many_runs(0.25,x,10)
simulate_many_runs(0.15,x,10,TRUE)
simulate_many_runs(0.35,x,10,FALSE,0.25,FALSE)
simulate_many_runs(0.25,x,10,FALSE,0.4,TRUE,0.08)
}
