% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_sir}
\alias{simulate_sir}
\title{Run simulation of infection spread until there are no longer any infected cells}
\usage{
simulate_sir(
  prob_infect = 0.25,
  input_matrix,
  plot = FALSE,
  seed = NULL
)
}
\arguments{
\item{prob_infect}{Probability of infection spreading}

\item{input_matrix}{Input matrix}

\item{plot}{Plot each simulation step}

\item{seed}{Optional random seed for reproducible simulation}
}
\value{
Results List
}
\description{
Run simulation of infection spread until there are no longer any infected cells
}
\examples{
x <- create_matrix(5,5)

simulate_sir(0.4,x)
}
