% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{simulate_sir}
\alias{simulate_sir}
\title{Run SIR simulation}
\usage{
simulate_sir(inf_prob = 0.5, row, col, start_inf = 0.125)
}
\arguments{
\item{inf_prob}{Probability of infection spreading}

\item{row}{Number of rows in matrix}

\item{col}{Number of columns in matrix}

\item{start_inf}{Probability of cells initially infected}
}
\value{
List of total steps taken, starting infection probability, infection spread probability, and final SIR matrix result.
}
\description{
Run SIR simulation
}
\examples{
simulate_sir(0.5, 20, 30, 0.125)
}
