% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSimulator.R
\name{simulate_sir}
\alias{simulate_sir}
\title{Run simulation of infection spread until there are no longer any infected cells}
\usage{
simulate_sir(
  prob_infect = 0.25,
  input_matrix,
  auto_immunity = FALSE,
  imm_prob = 0.5,
  allow_death = FALSE,
  fat_prob = 0.15
)
}
\arguments{
\item{prob_infect}{Probability of infection spreading}

\item{input_matrix}{Input matrix}

\item{auto_immunity}{Enable auto immunity after infection}

\item{imm_prob}{Probability of getting immunity}

\item{allow_death}{Enable death}

\item{fat_prob}{Probability of fatality}
}
\value{
Results List
}
\description{
Run simulation of infection spread until there are no longer any infected cells
}
\examples{
x <- create_matrix(5,5)

simulate_sir(0.4,x)
simulate_sir(0.15,x,TRUE)
simulate_sir(0.35,x,FALSE,0.5,FALSE)
simulate_sir(0.55,x,FALSE,0.35,TRUE,0.1)
}
